refactor(pipeline-transcription): types/ barrel + internal/ layout and canonical pipeline.test.ts
Move domain type modules (transcript-event, stage, lookup, pipeline-config) into
src/types/ (re-exported top-level to preserve consumers), move private asr util
into src/internal/, rename pipeline-integration.test.ts -> pipeline.test.ts.
Keeps the PipelineRuntime/Stage abstraction. plugin-transcription and
react-ui-transcription build unchanged.
