+++ packages/core/compute/pipeline-transcription/src/index.ts
@@ -2,10 +2,8 @@
 // Copyright 2026 DXOS.org
 //
 
-export * from './TranscriptEvent';
-export * from './Stage';
-export * from './lookup';
-export * from './PipelineConfig';
+// Domain types are re-exported top-level (not a `Type` namespace) because consumers import them directly.
+export * from './types';
 export * from './model-routing';
 export * from './PipelineRuntime';
 export * from './live';
