+++ packages/core/compute/pipeline-transcription/src/live.ts
@@ -10,7 +10,7 @@ import { EffectEx } from '@dxos/effect';
 import { type ContentBlock } from '@dxos/types';
 
 import { PipelineRuntime, type RunOptions } from './PipelineRuntime';
-import { TranscriptEvent } from './TranscriptEvent';
+import { TranscriptEvent } from './types';
 
 /**
  * Imperative handle to a running pipeline fed by a live ASR source. The source pushes events as they
