+++ packages/core/compute/pipeline-transcription/src/pipeline.test.ts
@@ -11,11 +11,11 @@ import { EchoTestBuilder } from '@dxos/echo-client/testing';
 import { EffectEx } from '@dxos/effect';
 import { type ContentBlock, Organization, Person } from '@dxos/types';
 
-import { makeDatabaseLookup } from './lookup';
+import { makeDatabaseLookup } from './types/lookup';
 import { PipelineRuntime } from './PipelineRuntime';
 import { type CommitFn } from './PipelineRuntime';
 import { makeCorrectionStage, makeExtractionStage } from './stages';
-import { TranscriptEvent } from './TranscriptEvent';
+import { TranscriptEvent } from './types/transcript-event';
 
 // Reproduces the Pipeline testbench path: correction + extraction through the runtime, asserting that
 // recognized entities are linked in the final corrected text (regardless of stage commit ordering).
