+++ packages/core/compute/pipeline-transcription/src/types/index.ts
@@ -1,0 +1,8 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './transcript-event';
+export * from './stage';
+export * from './lookup';
+export * from './pipeline-config';
