+++ packages/core/compute/pipeline-transcription/src/types/lookup.ts
