+++ packages/core/compute/pipeline-transcription/src/types/pipeline-config.ts
