+++ packages/core/compute/pipeline-transcription/src/types/stage.ts
