+++ packages/core/compute/pipeline-transcription/src/types/transcript-event.ts
